+++ client_requirements.txt
@@ -68,4 +68,4 @@ llama-cpp-python==v0.2.90
 faster-whisper==1.1.0
 psutil==6.1.0
 Pillow==10.2.0
-en-core-web-trf @ https://huggingface.co/spacy/en_core_web_trf/resolve/main/en_core_web_trf-any-py3-none-any.whl
+en_core_sci_md @ https://s3-us-west-2.amazonaws.com/ai2-s2-scispacy/releases/v0.5.4/en_core_sci_md-0.5.4.tar.gz
